// src/tests/api/api_testing_with_pom/types/api.types.ts
export interface Post {
  userId: number;
  id?: number;
  title: string;
  body: string;
}

export interface User {
  id?: number;
  name: string;
  username: string;
  email: string;
  address?: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone?: string;
  website?: string;
  company?: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}

export interface Comment {
  postId: number;
  id?: number;
  name: string;
  email: string;
  body: string;
}

export interface ApiResponse<T> {
  status: number;
  data: T;
  headers: Record<string, string>;
}